Handle circular static logger initialization in EnvironmentHelpers
DatadogLogging creates a shared logger instance. Creating this instance requires using `EnvironmentHelpers`.
This in turn tries to get the `DataDogLogging.SharedLogger`
This is initially set to a "null" logging implementation, and will remain as such for the lifetime of the program
Using a `Lazy<>` means the logger is only retrieved when ther's an exception in `EnvironmentHelpers`, hopefully after `SharedLogger` has been updated